main_menu/requirements.txt
pygame==2.5.2
pyserial==3.5
requests==2.31.0